requirements.txt
fastapi
uvicorn
python-telegram-bot
langchain
groq
python-dotenv